requirements.txt
coreapi==2.3.3
coreapi-cli==1.0.9
coverage==4.5.1
Django==2.1
djangorestframework==3.8.2
psycopg2-binary==2.7.5